refactor: extract PDF media detection
